requirements.txt
fastapi==0.104.1
mangum==0.17.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.1
pinecone-client==3.0.0
openai==1.6.1
pypdf==3.17.0
pydantic==2.5.0
python-multipart==0.0.6
